fix: align ActiveWeek entity id type with INT database column

+++ backend/src/main/java/edu/tcu/projectpulse/activeweek/ActiveWeekRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface ActiveWeekRepository extends JpaRepository<ActiveWeek, Long> {
+public interface ActiveWeekRepository extends JpaRepository<ActiveWeek, Integer> {
     List<ActiveWeek> findBySectionIdOrderByStartDateAsc(Long sectionId);
     void deleteBySectionId(Long sectionId);
     List<ActiveWeek> findBySectionIdAndActiveTrueAndStartDateLessThanEqualOrderByStartDateDesc(Long sectionId, LocalDate today);
